examples/metaflow/requirements.txt
metaflow
azure-identity
azure-batch
cfa-azure @ git+https://github.com/CDCgov/cfa_azure.git